# Revert removal of includes from scopes