For details/motivation please see #366