@msievers 2016-09-22
@pyromaniac Can you give this a review ?

I decided to leave the current implementation in place for ActiveRecord < 4, because its `pluck` cannot handle relation-based conditions anyway. For ActiveRecord 4 I handled a unconsistency between `ActiveRecord::Relation` and `ActiveRecord::Associations::CollectionProxy` and made the (sane) ActiveRecord > 4 code the default one.