@msievers 2016-09-22
Unfortunatly the `:active_record` tag isn't enough here, because the scope does use `.where.not` which only works in ActiveRecord >= 4, so I have to guard this even for the ActiveRecord gemfiles. But if I don't check that `ActiveRecord` is present with this `const_defined?`, the spec will fail with `uninitialized constant ActiveRecord` for the non-AR gemfiles.

The `::ActiveRecord::Associations::CollectionProxy` is given (only) from the specs inside `spec/chewy/type/observe_spec.rb`, without knowing why in detail. I only noticed this because of (some) failing specs with the ActiveRecord 4 gemfiles.

Regarding performance I could do some imports with some real datasets to get a feeling, but I think in doubt, the user is free to use some "optimized" scopes for chewy if performance should really be an issue.