@msievers 2016-09-22
Regarding performance I could do some imports with some real datasets to get a feeling, but I think in doubt, the user is free to use some "optimized" scopes for chewy if performance should really be an issue.