@msievers 2016-09-22
Ok, just give me a 1-2 days. Will see to get some numbers quickly.