@msievers 2016-09-23
@pyromaniac It seems you are right regarding the performance implications when using pluck with a scope, which utilizes `includes`. I build a basic benchmark to get some numbers.

https://github.com/msievers/chewy/tree/benchmark/pluck_with_includes

While other operations are smart enough to detect that the included relation isn't needed at all, pluck doesn't. (even with the latest ActiveRecord 5).

```ruby
class Country < ActiveRecord::Base
  has_many :cities

  scope :with_includes, -> { includes(:cities) }
end

class City < ActiveRecord::Base
  belongs_to :country
end
```

```ruby
Benchmark.ips do |x|
  x.report('with includes') do
    Country.with_includes.pluck(:id)
  end

  x.report('without includes') do
    Country.pluck(:id)
  end

  x.compare!
end
```

```
Warming up --------------------------------------
       with includes     1.000  i/100ms
    without includes     6.000  i/100ms
Calculating -------------------------------------
       with includes      3.601  (± 0.0%) i/s -     18.000  in   5.000153s
    without includes     68.887  (± 1.5%) i/s -    348.000  in   5.053070s

Comparison:
    without includes:       68.9 i/s
       with includes:        3.6 i/s - 19.13x  slower
```

Dataset is `10000` countries, with `10` cities each.

The performance drop is hugh and I don't think this is reasonable. So I would close this, unless anybody has objections.