@pyromaniac 2016-09-24
You are still able to use `includes().joins()` to bypass this behavior if you need. I've checked rails 5 and it behaves the same way as 4, no optimization there. Seems to be intended.

Anyway, I actually like how did you refactor rails 5 related module, it seems to be more accurate to me.